Refactor backend into app, routes, and controllers
